backend/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
idna==3.10
python-dotenv==1.1.1
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
pandas==2.2.3
openpyxl==3.1.5
Pillow==11.1.0
django-extensions==3.2.3
python-dateutil==2.9.0.post0